backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.11
alembic==1.13.2
celery==5.4.0
redis==5.0.8
httpx==0.27.2
yt-dlp==2026.3.3
python-dotenv==1.0.1
numpy>=1.26.4
python-docx==1.1.2
fpdf2==2.8.7
ffmpeg-python==0.2.0
opencv-python-headless==4.11.0.86
pytesseract==0.3.13
Pillow==12.0.0
sentence-transformers==3.3.1
setuptools>=68.0.0
faster-whisper==1.2.1